Split the code into View, Model and ViewModel files.
